feature: update pipe lineages and materializations (et/somenergia-jardiner!94)
* new: create an observability model directory and migrate some models

* change: add dbt tag for dset_responses_fresh

this tag accomodates for those models that are materialized every 5 minutes and the models that are materialized every hour

* change: smplify values incremental and remove time filters

make values incremental model depend on single model instead of union and remove time filter limiting the amount of data used

* change: update references in tests from union to single model only

* fix: bad linting

* change: reduce materialization frequency and avoid collisions

* chore: formatting

* dev: update test

* dev: add sqlfluff exception for first line indentation

* change: update test to use materialized model and different times

* dev: better pre-commit variable usage

* dev: let pre-commit at ci ignore sqlfluff linting and formatting

* chore: bump style using gitlab-ci

* dev: update pre-commit with sqlfluff-lint and update rules

* change: make spined model feed from union model

before it fed only from materialized model, now it feeds from the union which is the materialized and view models together

also bump format

* change: rename model from materialized_one_hour_late to just materialized

* change: remove time filter of one month at materialized table

because it is not needed. we will use the view to fetch the remaining data between one materialization cycle and the next.